Split Insert component
